refact(migration nominal): change from decimal to integer

+++ database/migrations/2025_10_22_072155_create_pemasukan_b_o_p_s_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('masuk_bop', function (Blueprint $table) {
             $table->id();
             $table->date('tgl');
-            $table->decimal('nominal', 12,2);
+            $table->integer('nominal');
             $table->text('ket');
             $table->string('bkt_nota');
             $table->timestamps();
